Publish prerelease yak package from ci build (#2)
Adds a manifest.yml file and a powershell script which sets the version,
zips up the package and pushes it to the yak server.

Patches AssemblyInformationalVersion and tweaks the GhCanvasViewportInfo
class to use the AssemblyInformationalVersion for its Version property.
Since AssemblyInfo patching is easy to do with appveyor, this seems like
a good way to inject a semantic version string (inc. prerelease
identifier and build number) into the version that is read by
Grasshopper. It's important for the Version property to be kept in sync
with the package version for package restore to function properly.

Downloads an 'ilmerged' release of yak.exe (v0.5) to do the pushing.

Deployments happen on the `master` branch only.

See YAK-127